Move views.py down a level
