+++ config/urls.py
@@ -1,12 +1,12 @@
 from django.conf.urls import url
 
-from profile import views as profile_views
+from profile.eig_apps import views as eig_apps_views
 
 
 urlpatterns = [
     url(
         r'^$',
-        profile_views.LandingPageView.as_view(),
+        eig_apps_views.LandingPageView.as_view(),
         name='index'
     ),
 ]
